fix build on windows
