+++ api/pkg/config/runner_config.go
@@ -29,7 +29,7 @@ type Runtimes struct {
 	V2Engine bool `envconfig:"RUNTIME_V2_ENGINE" default:"true"`
 	Axolotl  struct {
 		Enabled      bool          `envconfig:"RUNTIME_AXOLOTL_ENABLED" default:"true"`
-		WarmupModels []string      `envconfig:"RUNTIME_AXOLOTL_WARMUP_MODELS" default:"mistralai/Mistral-7B-Instruct-v0.1"`
+		WarmupModels []string      `envconfig:"RUNTIME_AXOLOTL_WARMUP_MODELS" default:""`
 		InstanceTTL  time.Duration `envconfig:"RUNTIME_AXOLOTL_INSTANCE_TTL" default:"10s"`
 	}
 	Ollama OllamaRuntimeConfig
@@ -37,6 +37,6 @@ type Runtimes struct {
 
 type OllamaRuntimeConfig struct {
 	Enabled      bool          `envconfig:"RUNTIME_OLLAMA_ENABLED" default:"true"`
-	WarmupModels []string      `envconfig:"RUNTIME_OLLAMA_WARMUP_MODELS" default:"llama3:instruct,llama3.1:8b-instruct-q8_0,llama3.2:1b-instruct-q8_0,llama3.2:3b-instruct-q8_0,phi3.5:3.8b-mini-instruct-q8_0"`
+	WarmupModels []string      `envconfig:"RUNTIME_OLLAMA_WARMUP_MODELS" default:"llama3.1:8b-instruct-q8_0"`
 	InstanceTTL  time.Duration `envconfig:"RUNTIME_OLLAMA_INSTANCE_TTL" default:"10s"`
 }
